split ViewerInfo and manual test fixes
