+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_models.py
@@ -1,0 +1,59 @@
+import logging
+import uuid
+
+from aiopg.sa.result import RowProxy
+from models_library.services import ServiceKey, ServiceVersion
+from pydantic import BaseModel, Field
+
+MEGABYTES = 1024 * 1024
+_BASE_UUID = uuid.UUID("ca2144da-eabb-4daf-a1df-a3682050e25f")
+
+
+logger = logging.getLogger(__name__)
+
+
+class ServiceInfo(BaseModel):
+    key: ServiceKey
+    version: ServiceVersion
+
+    label: str = Field(..., description="Display name")
+
+    is_guest_allowed: bool = True
+
+    @property
+    def footprint(self) -> str:
+        return f"{self.key}:{self.version}"
+
+    @property
+    def title(self) -> str:
+        """human readable title"""
+        return f"{self.label.capitalize()} v{self.version}"
+
+
+class ViewerInfo(ServiceInfo):
+    """
+    Here a viewer denotes a service
+      - that supports (i.e. can consume) a specific filetype and
+      - that is available to everyone
+
+    and therefore it can be dispatched to both guest and active users
+    to visualize a file of that type
+    """
+
+    filetype: str = Field(..., description="Filetype associated to this viewer")
+
+    input_port_key: str = Field(
+        ...,
+        description="Name of the connection port, since it is service-dependent",
+    )
+
+    @classmethod
+    def create_from_db(cls, row: RowProxy) -> "ViewerInfo":
+        return cls(
+            key=row["service_key"],
+            version=row["service_version"],
+            filetype=row["filetype"],
+            label=row["service_display_name"] or row["service_key"].split("/")[-1],
+            input_port_key=row["service_input_port"],
+            is_guest_allowed=row["is_guest_allowed"],
+        )
